Bugfix for DVC
This bugfix handles issues with dvc files. The regex has been extended to specifically look for output files that end with *.tsv ($ ensures this is the end of the string in the pattern). 

This resolves breaking with *.tsv.dvc files created when using data version control. Might be required in other cases too, will give it a try.